adding checking relations on option delete

+++ module/attribute/src/Infrastructure/Handler/Option/DeleteOptionCommandHandler.php
@@ -9,10 +9,12 @@ declare(strict_types = 1);
 
 namespace Ergonode\Attribute\Infrastructure\Handler\Option;
 
-use Ergonode\Attribute\Domain\Repository\OptionRepositoryInterface;
 use Ergonode\Attribute\Domain\Command\Option\DeleteOptionCommand;
-use Webmozart\Assert\Assert;
 use Ergonode\Attribute\Domain\Entity\AbstractOption;
+use Ergonode\Attribute\Domain\Repository\OptionRepositoryInterface;
+use Ergonode\Core\Infrastructure\Exception\ExistingRelationshipsException;
+use Ergonode\Core\Infrastructure\Resolver\RelationshipsResolverInterface;
+use Webmozart\Assert\Assert;
 
 /**
  */
@@ -24,11 +26,21 @@ class DeleteOptionCommandHandler
     private OptionRepositoryInterface $repository;
 
     /**
-     * @param OptionRepositoryInterface $repository
+     * @var RelationshipsResolverInterface
+     *
      */
-    public function __construct(OptionRepositoryInterface $repository)
-    {
+    private RelationshipsResolverInterface $relationshipsResolver;
+
+    /**
+     * @param OptionRepositoryInterface      $repository
+     * @param RelationshipsResolverInterface $relationshipsResolver
+     */
+    public function __construct(
+        OptionRepositoryInterface $repository,
+        RelationshipsResolverInterface $relationshipsResolver
+    ) {
         $this->repository = $repository;
+        $this->relationshipsResolver = $relationshipsResolver;
     }
 
     /**
@@ -44,6 +56,12 @@ class DeleteOptionCommandHandler
             AbstractOption::class,
             sprintf('Option with ID "%s" not found', $command->getId())
         );
+        $relationships = $this->relationshipsResolver->resolve($command->getId());
+
+        if (!$relationships->isEmpty()) {
+            throw new ExistingRelationshipsException($command->getId());
+        }
+
         $this->repository->delete($option);
     }
 }
